pyproject.toml
[project]
name = "nlp-at2"
version = "0.1.0"
description = "FinSight AI: Decision support system for financial document analysis"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    # Core Data & Analysis
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "openpyxl>=3.1.5",
    
    # Financial & SEC Data Ingestion
    "yfinance>=1.2.0",
    "sec-edgar-downloader>=5.1.0",
    "requests>=2.33.1",
    
    # NLP & Text Processing
    "langchain>=1.2.15",
    "langchain-core>=1.3.2",
    "langchain-text-splitters>=1.1.1",
    "nltk>=3.9.4",
    "pysbd>=0.3.4",
    "tokenizers>=0.22.2",
    "tiktoken>=0.7.0",           # Required for OpenAI token counting/splitting
    "bs4>=0.0.2",
    "beautifulsoup4>=4.14.3",
    "tqdm>=4.67.3",
    "openai>=2.32.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "sentence-transformers>=2.2.2",
    
    # Vector Database & RAG
    "chromadb>=1.5.5",
    "langchain-chroma>=1.1.0",
    "langchain-mcp-adapters", 
    "ragas>=0.4.3",
    "datasets>=2.18.0",          # Required for Ragas Evaluation (Hugging Face Datasets)
    "langchain-openai>=1.2.1",
    "langchain-community>=0.4.1",
    
    # Backend, Database & API
    "fastapi>=0.136.1",
    "uvicorn[standard]>=0.43.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "asyncpg>=0.31.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.9",   # Required for FastAPI form/file handling
    
    # Visualization & UI
    "streamlit>=1.32.0",         # Added for your FinSight AI frontend
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "ipywidgets>=8.1.8",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
    "notebook>=7.5.5",

    # Testing & Quality Assurance
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "pytest-timeout>=2.4.0",
    "httpx>=0.28.1",
    "respx>=0.23.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["backend", "RAG-Ingestion"]
asyncio_mode = "auto"
timeout = 120
markers = [
    "unit: Pure logic tests with no I/O",
    "integration: Wired components with mocked external services",
    "api: FastAPI contract and endpoint tests",
    "rag: Retrieval quality and RAGAS evaluation",
    "ingestion: Data pipeline and SEC processing tests",
    "live: Tests requiring real API keys (OpenAI, etc.)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["backend/app", "MCP", "RAG-Ingestion"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
show_missing = true
fail_under = 60